Making RestaurantMenuHeader fixed on top when scrolling
