+++ src/components/RestaurantMenuItem.js
@@ -1,6 +1,6 @@
 import {Text} from 'native-base';
-import {ActivityIndicator, StyleSheet} from 'react-native';
-import {TouchableOpacity} from 'react-native-gesture-handler';
+import React from 'react';
+import {StyleSheet} from 'react-native';
 
 const styles = StyleSheet.create({
   menuItem: {
@@ -8,6 +8,7 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
     paddingVertical: 10,
+    marginBottom: 500,
   },
   menuItemText: {
     fontSize: 16,
@@ -18,13 +19,13 @@ const RestaurantMenuItem = ({item, onPress, isLoading}) => {
   const enabled = item.hasOwnProperty('enabled') ? item.enabled : true;
 
   return (
-    //   <TouchableOpacity
-    //     style={styles.menuItem}
-    //     onPress={enabled ? () => onPress(item) : null}
-    //     testID={`menuItem:${item.sectionIndex}:${item.index}`}>
-    //     <Text style={styles.menuItemText}>{item.name}</Text>
-    //     {isLoading && <ActivityIndicator color="#c7c7c7" size="small" />}
-    //   </TouchableOpacity>
+    // <TouchableOpacity
+    //   style={styles.menuItem}
+    //   onPress={enabled ? () => onPress(item) : null}
+    //   testID={`menuItem:${item.sectionIndex}:${item.index}`}>
+    //   <Text style={styles.menuItemText}>{item.name}</Text>
+    //   {isLoading && <ActivityIndicator color="#c7c7c7" size="small" />}
+    // </TouchableOpacity>
     <Text>item</Text>
   );
 };
